requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.5.0
requests>=2.31.0 